fix postimg data types mismatch
also remove the "postimgflag_t" typedef since its useless
and use smoler type for everything we dont need much space lol

+++ src/p_mobj.c
@@ -3244,7 +3244,7 @@ void P_DestroyRobots(void)
 void P_CalcChasePostImg(player_t *player, camera_t *thiscam)
 {
 	const boolean flipcam = (player->pflags & PF_FLIPCAM && !(player->pflags & PF_NIGHTSMODE) && player->mo->eflags & MFE_VERTICALFLIP);
-	UINT16 postimgtype = 0;
+	UINT8 postimgtype = 0;
 
 	if (encoremode)
 		postimgtype |= POSTIMG_MIRROR;
